reorganize tests and gridding code
