+++ src/fftvis/utils.py
@@ -12,7 +12,6 @@ from .core.utils import (
     get_pos_reds,
     get_plane_to_xy_rotation_matrix,
     get_task_chunks,
-    check_antpos_griddability
 )
 
 
@@ -43,5 +42,4 @@ __all__ = [
     "get_plane_to_xy_rotation_matrix",
     "get_task_chunks",
     "inplace_rot",
-    "check_antpos_griddability",
 ]
